requirements.txt
streamlit>=1.34.0
chromadb>=0.5.0
openai>=1.30.0
tiktoken>=0.7.0
pypdf>=4.2.0
markdown-it-py>=3.0.0
python-dotenv>=1.0.1
rapidfuzz>=3.9.3
